fix: resolve clippy warning for redundant pattern matching

+++ src-tauri/src/obs/commands.rs
@@ -35,7 +35,7 @@ impl OBSCommands {
             .await
             .context("Failed to create scene item")?;
 
-        Ok(item_id as i64)
+        Ok(item_id)
     }
 
     pub async fn remove_scene_item(
